+++ src/spox/_public.py
@@ -2,7 +2,7 @@
 
 import contextlib
 import itertools
-from typing import Dict, Optional, Protocol
+from typing import Dict, List, Optional, Protocol
 
 import numpy as np
 import onnx
@@ -218,14 +218,6 @@ def inline(model: onnx.ModelProto) -> _InlineCall:
     _signature_msg = f"signature {in_names}{_defaults_msg} -> {out_names}"
 
     model = _copy_model(model)
-    # FIXME: Renaming does not work on subgraphs as of ONNX 1.13/1.14.
-    for node in model.graph.node:
-        for attr in node.attribute:
-            if attr.HasField("g") or attr.graphs:
-                raise ValueError(
-                    "Inlining models with subgraphs is not supported due to "
-                    "lack of upstream support for renaming values in subgraphs."
-                )
     # FIXME: Support for functions is a bit involved, as it interacts with build.
     if model.functions:
         raise ValueError(
@@ -240,6 +232,28 @@ def inline(model: onnx.ModelProto) -> _InlineCall:
         info.type.CopyFrom(
             _strip_dim_symbol(Type._from_onnx(info.type), lambda x: True)._to_onnx()
         )
+    # We handle everything related to initializers here, as currently build does not support them too well
+    # Overridable initializers are saved to in_defaults, non-overridable replaced with Constant
+    preamble: List[onnx.NodeProto] = []
+    input_names = {i.name for i in model.graph.input}
+    preamble.extend(
+        onnx.helper.make_node("Constant", [], [i.name], value=i)
+        for i in model.graph.initializer
+        if i.name not in input_names
+    )
+    preamble.extend(
+        onnx.helper.make_node("Constant", [], [i.values.name], sparse_value=i)
+        for i in model.graph.sparse_initializer
+        if i.values.name not in input_names
+    )
+    del model.graph.initializer[:]
+    del model.graph.sparse_initializer[:]
+    # The API on the protobuf list is a bit limited
+    # - this prepends the preamble before the rest of the nodes
+    model.graph.node.reverse()
+    model.graph.node.extend(reversed(preamble))
+    model.graph.node.reverse()
+    # Now we can assume the graph has no initializers
 
     def inline_inner(*args: Var, **kwargs: Var) -> Dict[str, Var]:
         for name, arg in zip(in_names, args):